Add files via upload
Shoutout to the Hololyzer dev(s) for an amazing site with a lot of data, they're the inspiration for this (alongside wanting more practice with Shiny) as well as the source of the superchats! I'd like to get more practice with scraping more sites with better packages like Selenium, so I'll try to update this in the future to scrape directly from YouTube, and also add plotly visualizations in the near future.